Move lmnet/lmnet/utils to blueoil/utils
