+++ blueoil/cmd/convert_weight_from_darknet.py
@@ -23,8 +23,8 @@ import tensorflow as tf
 from lmnet import environment
 from lmnet.networks.classification.darknet import Darknet
 from lmnet.networks.object_detection.yolo_v2 import YoloV2
-from lmnet.utils import config as config_util
-from lmnet.utils import executor
+from blueoil.utils import config as config_util
+from blueoil.utils import executor
 
 
 def convert(config, weight_file):
