+++ blueoil/cmd/evaluate.py
@@ -25,9 +25,9 @@ from lmnet import environment
 from lmnet.datasets.base import ObjectDetectionBase
 from lmnet.datasets.dataset_iterator import DatasetIterator
 from lmnet.datasets.tfds import TFDSClassification, TFDSObjectDetection
-from lmnet.utils import config as config_util
-from lmnet.utils import executor, module_loader
-from lmnet.utils.predict_output.writer import save_json
+from blueoil.utils import config as config_util
+from blueoil.utils import executor, module_loader
+from blueoil.utils.predict_output.writer import save_json
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
