+++ blueoil/cmd/train.py
@@ -21,7 +21,7 @@ import yaml
 
 from blueoil.generate_lmnet_config import generate
 from executor.train import run as run_train
-from lmnet.utils import horovod as horovod_util
+from blueoil.utils import horovod as horovod_util
 
 
 def run(blueoil_config_file, experiment_id):
