+++ blueoil/utils/image.py
