+++ blueoil/utils/random.py
