+++ lmnet/lmnet/data_augmentor.py
@@ -22,7 +22,7 @@ import numpy as np
 from PIL import Image, ImageEnhance, ImageFilter
 
 from lmnet import data_processor, pre_processor
-from lmnet.utils.box import fill_dummy_boxes, crop_boxes, iou
+from blueoil.utils.box import fill_dummy_boxes, crop_boxes, iou
 
 
 class Blur(data_processor.Processor):
