+++ lmnet/lmnet/datasets/bdd100k.py
@@ -7,7 +7,7 @@ import os
 import numpy as np
 
 from lmnet.datasets.base import ObjectDetectionBase, SegmentationBase
-from lmnet.utils.image import load_image
+from blueoil.utils.image import load_image
 
 
 class BDD100KObjectDetection(ObjectDetectionBase):
