+++ lmnet/lmnet/datasets/camvid.py
@@ -22,7 +22,7 @@ from PIL import Image
 
 from lmnet.common import get_color_map
 from lmnet.datasets.base import SegmentationBase, StoragePathCustomizable
-from lmnet.utils.random import shuffle, train_test_split
+from blueoil.utils.random import shuffle, train_test_split
 
 
 def get_image(filename, convert_rgb=True, ignore_class_idx=None):
