+++ lmnet/lmnet/datasets/cifar100.py
@@ -21,7 +21,7 @@ import numpy as np
 
 from lmnet import data_processor
 from lmnet.datasets.base import Base
-from lmnet.utils.random import shuffle
+from blueoil.utils.random import shuffle
 
 
 class Cifar100(Base):
