+++ lmnet/lmnet/datasets/ilsvrc_2012.py
@@ -21,7 +21,7 @@ import numpy as np
 import pandas as pd
 
 from lmnet import data_processor
-from lmnet.utils.image import load_image
+from blueoil.utils.image import load_image
 from lmnet.datasets.base import Base
