+++ lmnet/lmnet/datasets/pascalvoc_base.py
@@ -21,7 +21,7 @@ import xml.etree.ElementTree as ET
 import numpy as np
 import pandas as pd
 
-from lmnet.utils.image import load_image
+from blueoil.utils.image import load_image
 from lmnet.datasets.base import ObjectDetectionBase
