+++ lmnet/lmnet/datasets/tfds.py
@@ -20,8 +20,8 @@ import tensorflow as tf
 import tensorflow_datasets as tfds
 
 from lmnet.datasets.base import Base, ObjectDetectionBase
-from lmnet.utils.tfds_builders.classification import ClassificationBuilder
-from lmnet.utils.tfds_builders.object_detection import ObjectDetectionBuilder
+from blueoil.utils.tfds_builders.classification import ClassificationBuilder
+from blueoil.utils.tfds_builders.object_detection import ObjectDetectionBuilder
 
 
 def _grayscale_to_rgb(record):
