+++ lmnet/lmnet/datasets/ytfaces.py
@@ -17,7 +17,7 @@ import os.path
 
 import numpy as np
 
-from lmnet.utils.image import load_image
+from blueoil.utils.image import load_image
 from lmnet.datasets.base import KeypointDetectionBase
