+++ lmnet/tests/executor_tests/test_build_tfds.py
@@ -21,7 +21,7 @@ from executor.train import run as train_run
 from lmnet import environment
 from lmnet.datasets.dataset_iterator import DatasetIterator
 from lmnet.datasets.tfds import TFDSClassification, TFDSObjectDetection
-from lmnet.utils import config as config_util
+from blueoil.utils import config as config_util
 
 _RUN_AS_A_SCRIPT = False
