+++ lmnet/tests/lmnet_tests/networks_tests/classification_test/test_lm_resnet_quantize.py
@@ -24,7 +24,7 @@ from lmnet.datasets.image_folder import ImageFolderBase
 from lmnet.networks.classification.lm_resnet import LmResnetQuantize
 from lmnet.pre_processor import Resize
 from blueoil.nn.quantizations import binary_mean_scaling_quantizer, linear_mid_tread_half_quantizer
-from lmnet.utils.executor import prepare_dirs
+from blueoil.utils.executor import prepare_dirs
 
 # Apply reset_default_graph() in conftest.py to all tests in this file.
 # Set test environment
