+++ lmnet/tests/lmnet_tests/networks_tests/keypoint_detection_tests/test_lm_single_pose_v1.py
@@ -21,7 +21,7 @@ from lmnet import environment
 from lmnet.common import Tasks
 from lmnet.datasets.mscoco_2017 import MscocoSinglePersonKeypoints
 from lmnet.networks.keypoint_detection.lm_single_pose_v1 import LmSinglePoseV1Quantize
-from lmnet.utils.executor import prepare_dirs
+from blueoil.utils.executor import prepare_dirs
 from lmnet.pre_processor import (
     DivideBy255,
     ResizeWithJoints,
