+++ lmnet/tests/lmnet_tests/networks_tests/object_detection_tests/test_yolo_v1.py
@@ -24,7 +24,7 @@ from lmnet.common import Tasks
 from lmnet.datasets.pascalvoc_2007 import Pascalvoc2007
 from lmnet.networks.object_detection.yolo_v1 import YoloV1
 from lmnet.pre_processor import ResizeWithGtBoxes
-from lmnet.utils.executor import prepare_dirs
+from blueoil.utils.executor import prepare_dirs
 
 # Apply reset_default_graph() in conftest.py to all tests in this file.
 # Set test environment
