+++ lmnet/tests/lmnet_tests/networks_tests/object_detection_tests/test_yolo_v2.py
@@ -26,7 +26,7 @@ from lmnet.datasets.pascalvoc_2007 import Pascalvoc2007
 from lmnet.networks.object_detection.yolo_v2 import YoloV2
 from lmnet.post_processor import NMS, ExcludeLowScoreBox, FormatYoloV2
 from lmnet.pre_processor import ResizeWithGtBoxes
-from lmnet.utils.executor import prepare_dirs
+from blueoil.utils.executor import prepare_dirs
 
 # Apply reset_default_graph() in conftest.py to all tests in this file.
 # Set test environment
