+++ lmnet/tests/lmnet_tests/networks_tests/object_detection_tests/test_yolo_v2_quantize.py
@@ -24,7 +24,7 @@ from lmnet.datasets.pascalvoc_2007 import Pascalvoc2007
 from lmnet.networks.object_detection.yolo_v2_quantize import YoloV2Quantize
 from lmnet.pre_processor import ResizeWithGtBoxes
 from blueoil.nn.quantizations import binary_channel_wise_mean_scaling_quantizer, linear_mid_tread_half_quantizer
-from lmnet.utils.executor import prepare_dirs
+from blueoil.utils.executor import prepare_dirs
 
 # Apply reset_default_graph() in conftest.py to all tests in this file.
 # Set test environment
