+++ lmnet/tests/lmnet_tests/networks_tests/segmentation_tests/test_lm_bisenet.py
@@ -26,7 +26,7 @@ from lmnet.datasets.camvid import Camvid
 from lmnet.networks.segmentation.lm_bisenet import LMBiSeNet
 from lmnet.post_processor import Bilinear, Softmax
 from lmnet.pre_processor import Resize
-from lmnet.utils.executor import prepare_dirs
+from blueoil.utils.executor import prepare_dirs
 
 # Apply reset_default_graph() and set_test_environment() in conftest.py to all tests in this file.
 pytestmark = pytest.mark.usefixtures("reset_default_graph", "set_test_environment")
