+++ lmnet/tests/test_configs.py
@@ -19,7 +19,7 @@ import os
 import pytest
 
 from lmnet import environment
-from lmnet.utils.config import _load_py, check_config, save_yaml
+from blueoil.utils.config import _load_py, check_config, save_yaml
 
 pytestmark = pytest.mark.usefixtures("set_test_environment")
