+++ output_template/python/run.py
@@ -22,10 +22,10 @@ import time
 
 from lmnet.nnlib import NNLib as NNLib
 
-from lmnet.utils.image import load_image
+from blueoil.utils.image import load_image
 from lmnet.common import Tasks
-from lmnet.utils.output import JsonOutput, ImageFromJson
-from lmnet.utils.config import (
+from blueoil.utils.output import JsonOutput, ImageFromJson
+from blueoil.utils.config import (
     load_yaml,
     build_pre_process,
     build_post_process,
