+++ output_template/python/usb_camera_demo.py
@@ -31,12 +31,12 @@ import numpy as np
 
 from lmnet.common import get_color_map
 from lmnet.nnlib import NNLib
-from lmnet.utils.config import (
+from blueoil.utils.config import (
     load_yaml,
     build_pre_process,
     build_post_process,
 )
-from lmnet.utils.demo import (
+from blueoil.utils.demo import (
     add_rectangle,
     add_fps,
     run_inference,
